Update the rest of the code to accommodate the change
