+++ apps/channels/tests/test_api_integration.py
@@ -4,7 +4,7 @@ import pytest
 from django.urls import reverse
 
 from apps.channels.models import ChannelPlatform, ExperimentChannel
-from apps.experiments.models import Participant
+from apps.participants.models import Participant
 from apps.utils.factories.experiment import ExperimentFactory
 from apps.utils.factories.team import TeamWithUsersFactory
 from apps.utils.tests.clients import ApiTestClient
