+++ templates/participants/partials/participant_experiments.html
@@ -22,7 +22,7 @@
         </tr>
         <tr x-show="experimentsRow === {{ forloop.counter }}" x-cloak>
           <td colspan="100%" class="px-3 py-2">
-            {% if perms.experiments.change_participantdata %}
+            {% if perms.participants.change_participantdata %}
               <div role="tablist" class="tabs tabs-bordered">
                 <input type="radio" name="experiment_tabs-{{ experiment.id }}" role="tab" class="tab" aria-label="Sessions" checked />
                 <div role="tabpanel" class="tab-content">
